chore: add test cases and separate interaction tests
